requirements.txt
asn1crypto==0.24.0
boto3==1.9.82
botocore==1.12.82
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.2
django-cors-headers==2.5.2
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
django-storages==1.7.1
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
future==0.17.1
httpie==0.9.9
idna==2.7
inflect==2.1.0
jmespath==0.9.3
pbr==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.5
python-decouple==3.1
python-http-client==3.1.0
pytz==2018.5
requests==2.22.0
s3transfer==0.1.13
six==1.12.0
sqlparse==0.2.4
stripe==2.31.0
Unipath==1.1
urllib3==1.23
